requirements.txt
python-telegram-bot==13.15
nmap
apscheduler
urllib3==1.26.6
python-nmap
